Finish up all the basic tests
