+++ lib/components/LogInOut.jsx
@@ -30,3 +30,5 @@ export const LogInOut = ({ user }) => {
     );
   }
 };
+
+module.exports = LogInOut;
