+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "webpack-dev-server",
     "build": "webpack",
-    "test": "mocha --compilers js:babel-core/register"
+    "test": "mocha --require test/helpers/setup.js"
   },
   "repository": {
     "type": "git",
